1.RailDraw修改一个复制错误；2.RailView天车显示测试代码完成，开始测试

+++ src/RailView/RailView/CodingRailCoordinates.cs
@@ -76,7 +76,7 @@ namespace RailView
             Point endPt = Point.Empty;
             bool isSearchedNextEle = false;
             startPt = GetPointOfInitFun(tempList, 0);
-            for (int i = 0; i < numOfList-1; i++)
+            for (int i = 0; i < numOfList; i++)
             {
                 switch (tempList[i].graphType)
                 {
@@ -170,11 +170,11 @@ namespace RailView
                         break;
                     }                  
                 }
-                if (!isSearchedNextEle)
+                if (!isSearchedNextEle && !(remailEleList.Count==0))
                 {
                     tempList.RemoveRange(i + 1, numOfList - i - 1);
                     tempList.AddRange(InitEleList(remailEleList));
-                    GetPointOfInitFun(tempList, i + 1);
+                    startPt= GetPointOfInitFun(tempList, i + 1);
                     remailEleList.Remove(tempList[i+1]);
                 }
                 isSearchedNextEle = false;
@@ -194,7 +194,7 @@ namespace RailView
                 case 1:
                     strTemp = (StraightEle)tempList[section];
                     returnPt = strTemp.startPoint;
-                    if ((strTemp.rotateAngle + 180) % 180 == 0)
+                    if (Math.Abs(strTemp.pointList[0].Y-strTemp.pointList[1].Y)<3)
                     {
                         if (strTemp.startPoint.X < strTemp.endPoint.X)
                             returnPt.X = strTemp.startPoint.X + offsetTemp;
